remove extra code for native
