+++ src/components/auth/torus/hooks/useSDKOptions.native.js
@@ -1,0 +1,8 @@
+import { useEffect } from 'react'
+import { noop } from 'lodash'
+
+export default (onOptions = noop) => {
+  useEffect(() => {
+    onOptions({ uxMode: 'popup' })
+  }, [])
+}
