+++ src/components/auth/torus/hooks/useTorus.js
@@ -1,18 +1,13 @@
-import { useEffect, useRef, useState } from 'react'
-import { get, noop } from 'lodash'
+import { useCallback, useEffect, useRef, useState } from 'react'
+import { noop } from 'lodash'
 
 import useMountedState from '../../../../lib/hooks/useMountedState'
-import createABTesting from '../../../../lib/hooks/useABTesting'
-import { DetectWebview, isMobileNative } from '../../../../lib/utils/platform'
 
 import TorusSDK from '../sdk/TorusSDK'
 import logger from '../../../../lib/logger/pino-logger'
+import useSDKOptions from './useSDKOptions'
 
 const log = logger.child({ from: 'AuthTorus' })
-
-const { useABTesting } = createABTesting('torusUxMode')
-
-const skipInitializeABTesting = isMobileNative
 
 export default (onInitialized = noop) => {
   const [sdk, setSDK] = useState()
@@ -20,23 +15,9 @@ export default (onInitialized = noop) => {
   const onInitializedRef = useRef(onInitialized)
   const mountedState = useMountedState()
 
-  // for native we need to skip initialize AB testing in this case
-  const [, abVariant, abTestInitialized] = useABTesting(null, null, null, skipInitializeABTesting)
+  const applySDKOptions = useCallback(options => setSDK(TorusSDK.factory(options)), [setSDK])
 
-  useEffect(() => {
-    if (abTestInitialized) {
-      const webview = new DetectWebview(get(global, 'navigator.userAgent'))
-      const isFacebookWebview = ['facebook', 'messenger'].includes(webview.browser)
-
-      log.debug('abTesting:', { abVariant, isFacebookWebview, skipInitializeABTesting })
-
-      // dont allow popup mode on facebook webview at all, since it doesnt work.
-      // allow popup mode only for native.
-      const torusUxMode = skipInitializeABTesting || (!isFacebookWebview && abVariant === 'A') ? 'popup' : 'redirect'
-
-      setSDK(TorusSDK.factory({ uxMode: torusUxMode }))
-    }
-  }, [abTestInitialized])
+  useSDKOptions(applySDKOptions)
 
   useEffect(() => {
     onInitializedRef.current = onInitialized
@@ -60,6 +41,7 @@ export default (onInitialized = noop) => {
         log.error('failed initializing torus', message, exception)
       }
     }
+
     if (sdk) {
       registerTorusWorker()
     }
